Rename merge_sim to execution_engine_integration
