+++ Cargo.toml
@@ -74,7 +74,7 @@ members = [
 
     "testing/ef_tests",
     "testing/eth1_test_rig",
-    "testing/merge_sim",
+    "testing/execution_engine_integration",
     "testing/node_test_rig",
     "testing/simulator",
     "testing/test-test_logger",
