+++ testing/execution_engine_integration/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "merge_sim"
+name = "execution_engine_integration"
 version = "0.1.0"
 edition = "2021"
